Move user settings into slice
